requirements.txt
flask_sqlalchemy
flask_migrate